backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydicom>=2.4.0
Pillow>=10.0.0
numpy>=1.26.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
scikit-image>=0.22.0
scipy>=1.12.0
matplotlib>=3.8.0
